+++ libs/physics-engine/physics-engine.vcxproj.filters
@@ -18,6 +18,12 @@
     <ClCompile Include="src\phy_engine.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
+    <ClCompile Include="src\math.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="src\plane.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\phy_engine.hpp">
